Add lcr logic to destroy and create tests for destroy action

+++ providers/record.rb
@@ -77,13 +77,18 @@ rescue Excon::Errors::UnprocessableEntity
 end
 
 def delete_record
- all_records_in_zone do |r|
-    if (( r.name == new_resource.name ) && ( r.type == new_resource.type ))
-      r.destroy
-      new_resource.updated_by_last_action(true)
-      Chef::Log.info "DNSimple: destroyed #{new_resource.type} record " +
-        "for #{new_resource.name}.#{new_resource.domain}"
+  if [:same, :different].include?(@current_resource.exists)
+    all_records_in_zone do |r|
+      if (( r.name == new_resource.name ) && ( r.type == new_resource.type ))
+        r.destroy
+        new_resource.updated_by_last_action(true)
+        Chef::Log.info "DNSimple: destroyed #{new_resource.type} record " +
+          "for #{new_resource.name}.#{new_resource.domain}"
+      end
     end
+  else
+    Chef::Log.info "DNSimple: no record found #{new_resource.name}.#{new_resource.domain}" +
+      " with type #{new_resource.type}"
   end
 end
